+++ R/RcppExports.R
@@ -121,6 +121,10 @@ cpp_dyn_abun_init_loglik <- function(C, eta_lambda, use_nb = FALSE, eta_logr = 0
     .Call(`_tulpaObs_cpp_dyn_abun_init_loglik`, C, eta_lambda, use_nb, eta_logr, deriv)
 }
 
+cpp_dyn_abun_p_loglik <- function(y, n_sites, T, J, K, site, eta_lambda, eta_p, eta_omega, eta_gamma, use_nb = FALSE, eta_logr = 0.0, deriv = TRUE) {
+    .Call(`_tulpaObs_cpp_dyn_abun_p_loglik`, y, n_sites, T, J, K, site, eta_lambda, eta_p, eta_omega, eta_gamma, use_nb, eta_logr, deriv)
+}
+
 cpp_dyn_abun_nuts_joint_logpost <- function(spec, theta, sigma_beta) {
     .Call(`_tulpaObs_cpp_dyn_abun_nuts_joint_logpost`, spec, theta, sigma_beta)
 }